Reimplement some tricky parts in C++11
With the arrival of MSVC 12.0 some amazing C++11 features not supported
before appear. More concisely, enum classes and variadic templates are
now supported.

This change reimplements some parts of the code using enum classes and
variadic templates where appropriate.

+++ oacsd/liboac-commons/src/logging.cpp
@@ -36,7 +36,7 @@ std::shared_ptr<logger> main_logger;
 
 const char*
 logger::level_str(log_level level)
-{ return LEVEL_STR[level]; }
+{ return LEVEL_STR[static_cast<int>(level)]; }
 
 std::string
 logger::get_time()
